Split file_editor.cpp into smaller files
